backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
    "mysqlclient>=2.2.8",
    "requests>=2.32.5",
]
